generalize translations route constraint
